Add Invoke-EditorServicesProcess function (#2)
* Add Invoke-EditorServicesProcess function

This change adds the Invoke-EditorServicesProcess function.  With this
function you directly invoke PowerShell code in the Integrated Console
from standard PowerShell.

Unlike the Enter-EditorServicesProcess, this function can be used in
a non-interactive context. Additionally it will invoke directly in
the default runspace, rather than a proxy runspace for the process.

- Fixed an unhandled exception that occurred when VSCode was not open.

- Moved most of the new runspace logic out of Enter-EditorServicesProcess
  and into a private function for maintainability.

* Update release notes